Updated for increased compatibility
There's code, which worked in older versions of ghostscript, for handling plus and minus Infinty and NaN, which were realtype objects. Newer versions of ghostscript are more compatible with Adobe, which produces undefinedresult for anything that might produce an Infinity. To handle this, we had added some code for finding the maximum real, which didn't work in older versions of ghostscript. That has now been fixed. It's unfortunate that there now seems to be no way to access in ghostscript NaN and +- Infinity that IEEE hardware provides. (And to think I was the one that reported this (now former) incompatibility.)